fix : 전체 URL을 기준으로 토큰 유효성 검사 실시

+++ src/main/java/com/example/tuktuk/schedule/service/ScheduleService.java
@@ -159,7 +159,7 @@ public class ScheduleService {
                   ScheduleReadResponseDto.from(schedule, hourlyRentFee))
               .toList());
       totalPage += schedulePage.getTotalPages();
-      totalElements += schedulePage.getTotalElements();
+      totalElements += (int) schedulePage.getTotalElements();
     }
 
     return SchedulePerStadiumResDto
